requirements.txt
azure-functions
azure-functions-durable
Pillow  # if using for image metadata
pyodbc
